Error fix caused by Kraken's naming system for asset pairs
When using Kraken's REST API together with pykrakenapi, some trading pairs (such as the very popular XBTUSD) produce a KeyError, even though the pair exists in Kraken's Asset Pair endpoint. This issue is a result of Kraken using a different name for the trading pair. In the case of XBTUSD, this is XXBTZUSD. This can be illustrated by printing the raw results from query_public. 
I propose a simple fix by resetting the pair value to that of the first key of the dictionary that's under the result key.

A more detailed explanation can be found here: https://stackoverflow.com/questions/63567624